authorize added

+++ DDBus/Controllers/AccountController.cs
@@ -6,12 +6,17 @@ using DDBus.Service;
 using System.Text.RegularExpressions;
 using System.Security.Claims;
 
+
+
+
+
 namespace Reflectly.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
     public class AccountController : ControllerBase
     {
+
         private readonly IConfiguration _configuration;
         private readonly CRUD_Service<Account> _Account_Service;
         public AccountController(
@@ -22,17 +27,24 @@ namespace Reflectly.Controllers
             _Account_Service = Account_Service;
         }
 
-        //[Authorize]
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> Get()
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if ((await is_root(userId!))) return BadRequest("permission denied");
             return Ok(await _Account_Service.GetAllAsync());
         }
 
-        //[Authorize]
+
+        [Authorize]
         [HttpGet("{id:length(24)}")]
         public async Task<IActionResult> GetAccountById(string id)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if ((await is_root(userId!))) return BadRequest("Không có quyền xóa");
+
+
             var account = await _Account_Service.GetByIdAsync(id);
 
             if (account == null)
@@ -44,27 +56,31 @@ namespace Reflectly.Controllers
             return Ok(account);
         }
 
-        //[Authorize]
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> CreateAccount([FromBody] Account newAccount)
         {
+
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if ((await is_root(userId!))) return BadRequest("Không có quyền xóa");
+
+
             if (newAccount == null)
             {
-                return BadRequest("Dữ liệu tài khoản không hợp lệ");
+                return BadRequest("Invalid account data");
             }
 
             if (!Regex.IsMatch(newAccount.Email, @"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$") || newAccount.Password.Length < 8)
-                return BadRequest("Tài khoản không hợp lệ");
+                return BadRequest("Account Invalid");
 
             var existingUser = await _Account_Service.Get_One_Item_Async(newAccount.Email, "Email");
             if (existingUser != null)
             {
-                return BadRequest("Email đã tồn tại");
+                return BadRequest("Email Existed.");
             }
 
             var hashedPassword = BCrypt.Net.BCrypt.HashPassword(newAccount.Password);
 
-            // Tạo đối tượng Account
             var accountToCreate = new Account
             {
                 Username = newAccount.Username,
@@ -79,32 +95,40 @@ namespace Reflectly.Controllers
             return Ok();
         }
 
-        //[Authorize]
+
+        [Authorize]
         [HttpPut("{id:length(24)}")]
         public async Task<IActionResult> UpdateAccount(string id, [FromBody] Account updatedAccount)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if ((await is_root(userId!))) return BadRequest("Không có quyền xóa");
+
             var existingAccount = await _Account_Service.GetByIdAsync(id);
             if (existingAccount == null)
             {
-                return NotFound(new { Message = "Không tìm thấy tài khoản" });
+                return NotFound(new { Message = "Account not found" });
             }
 
             await _Account_Service.UpdateAsync(id, updatedAccount);
             return Ok();
         }
 
-        //[Authorize]
+        [Authorize]
         [HttpDelete("{id:length(24)}")]
         public async Task<IActionResult> DeleteAccount(string id)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if ((await is_root(userId!))) return BadRequest("Không có quyền xóa");
+
             var existingAccount = await _Account_Service.GetByIdAsync(id);
             if (existingAccount == null)
             {
-                return NotFound(new { Message = "Không tìm thấy tài khoản" });
+                return NotFound(new { Message = "Account not found" });
             }
             await _Account_Service.DeleteAsync(id);
             return Ok();
         }
+
 
         private async Task<bool> is_root(string userid)
         {
@@ -112,5 +136,11 @@ namespace Reflectly.Controllers
             if (ac != null && ac.role == 0) return true;
             return false;
         }
+
+
     }
+
+
+
+
 }
